+++ index.js
@@ -55,7 +55,7 @@ function auth(req) {
   }
 
   // get header
-  var header = (req.req || req).headers.authorization
+  var header = getAuthorization(req.req || req)
 
   // parse header
   var match = credentialsRegExp.exec(header || '')
@@ -82,6 +82,19 @@ function auth(req) {
 
 function decodeBase64(str) {
   return new Buffer(str, 'base64').toString()
+}
+
+/**
+ * Get the Authorization header from request object.
+ * @private
+ */
+
+function getAuthorization(req) {
+  if (!req.headers || typeof req.headers !== 'object') {
+    throw new TypeError('argument req is required to have headers property')
+  }
+
+  return req.headers.authorization
 }
 
 /**