+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Improve error message when `req` argument is not an object
+  * Improve error message when `req` missing `headers` property
 
 1.0.3 / 2015-07-01
 ==================